add TS rules and linting
