--- frontend/components/navigation/navbar.jsx
@@ -1,12 +1,0 @@
-import { ConnectButton } from "@rainbow-me/rainbowkit";
-import styles from "../../styles/Navbar.module.css";
-export default function Navbar() {
-	return (
-		<nav className={styles.navbar}>
-			<a href="https://alchemy.com/?a=create-web3-dapp" target={"_blank"}>
-				<img className={styles.alchemy_logo} src="/cw3d-logo.png"></img>
-			</a>
-			<ConnectButton></ConnectButton>
-		</nav>
-	);
-}
