--- frontend/layout/mainLayout.jsx
@@ -1,10 +1,0 @@
-import Navbar from "../components/navigation/navbar";
-
-export default function MainLayout({ children }) {
-	return (
-		<div>
-            <Navbar />
-            {children}
-		</div>
-	);
-}
