+++ frontend/layout/mainLayout.tsx
@@ -1,0 +1,12 @@
+import Navbar from "../components/navigation/navbar"
+
+const MainLayout = ({ children }) => {
+  return (
+    <div>
+      <Navbar />
+      {children}
+    </div>
+  )
+}
+
+export default MainLayout
