+++ frontend/package.json
@@ -10,11 +10,11 @@
     "marketplace": "npx create-web3-dapp marketplace"
   },
   "dependencies": {
+    "@rainbow-me/rainbowkit": "^0.12.4",
+    "alchemy-sdk": "^2.6.2",
     "next": "13.1.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "alchemy-sdk": "^2.6.2",
-    "@rainbow-me/rainbowkit": "^0.12.4",
     "wagmi": "^0.12.6"
   },
   "devDependencies": {
@@ -32,6 +32,7 @@
     "@types/chai": "^4.3.3",
     "@types/mocha": "^10.0.0",
     "@types/node": "^18.11.3",
+    "@types/react": "^18.2.6",
     "chai": "^4.2.0",
     "dotenv": "^16.0.3",
     "ethers": "^5.7.2",
@@ -40,9 +41,6 @@
     "hardhat-deploy-ethers": "^0.3.0-beta.13",
     "hardhat-gas-reporter": "^1.0.9",
     "prettier": "^2.7.1",
-    "prettier-plugin-solidity": "^1.0.0-beta.24",
-    "solhint": "^3.3.7",
-    "solidity-coverage": "^0.8.2",
     "ts-node": "^10.9.1",
     "typechain": "^8.1.0",
     "typescript": "^4.8.4"
