--- frontend/pages/index.jsx
@@ -1,12 +1,0 @@
-import styles from "../styles/Home.module.css";
-import InstructionsComponent from "../components/InstructionsComponent";
-
-export default function Home() {
-  return (
-    <div>
-      <main className={styles.main}>
-        <InstructionsComponent></InstructionsComponent>
-      </main>
-    </div>
-  );
-}
